make searchResults a component
