+++ app/components/search-results/template.html
@@ -1,4 +1,4 @@
-<div class="container" ng-controller="SearchResultsCtrl">
+<div class="container">
   <div ng-if="search.documents !== undefined">
     <h2 ng-if="!search.documents.length">No results for <em>{{search.query}}</em></h2>
     <div ng-if="search.documents.length">
